requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
SQLAlchemy==2.0.37
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
